Extend io customization support: whole obj rule in split obj.
Extend the support for a rule that applies to several data members (and thus currently is applied as the 'object level') to
the case of a split object embedded within a split subobject of an object (previous it worked only for the collection case).

This fixes issue #???, see also https://cdcvs.fnal.gov/redmine/issues/25893.